+++ duckql_django/wrapper.py
@@ -1,19 +1,23 @@
 import copy
 import json
-from typing import Dict, List
+from typing import Dict, List, Type
 
+from django.contrib.auth.models import PermissionsMixin
 from django.db import connection
 from duckql import Query, Count, Constant
 
+from . import Schema
+
 
 class QueryWrapper:
-    def __init__(self, query: str):
+    def __init__(self, query: str, base_model: Type, user: PermissionsMixin = None):
         self._query = Query.parse_raw(query)
+        self._schema = Schema(base_model, user)
 
     @classmethod
-    def from_dict(cls, payload: Dict):
+    def from_dict(cls, payload: Dict, base_model: Type, user: PermissionsMixin = None):
         # So dirty, so nasty, so sad
-        return QueryWrapper(json.dumps(payload))
+        return QueryWrapper(json.dumps(payload), base_model, user)
 
     @staticmethod
     def _execute_query(query: Query) -> List[Dict]: